Move hasPendingTransaction check into initTransaction() and setActive()
This takes care of read, write and batch nested transactions.
